Split scrollbars into standalone primitives
