+++ composeunstyled-scroll-area/src/commonMain/kotlin/com/composeunstyled/ScrollArea.kt
@@ -38,12 +38,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.CoroutineStart
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.sync.Mutex
-import kotlinx.coroutines.sync.withLock
 import kotlin.js.JsName
 import kotlin.math.abs
 import kotlin.math.roundToInt
@@ -73,8 +67,8 @@ fun ScrollArea(
 ) {
   Box(modifier) {
     val boxScope = this
-    val scrollAreaScope = remember {
-      ScrollAreaScope(boxScope, state)
+    val scrollAreaScope = remember(boxScope) {
+      ScrollAreaScope(boxScope)
     }
     scrollAreaScope.content()
   }
@@ -82,7 +76,6 @@ fun ScrollArea(
 
 class ScrollAreaScope internal constructor(
   private val boxScope: BoxScope,
-  internal val scrollAreaState: ScrollAreaState,
 ) {
   fun Modifier.align(alignment: Alignment): Modifier {
     return with(boxScope) {
@@ -378,76 +371,4 @@ internal class LazyGridScrollAreaScrollAreaState(
   }
 
   override val lineSpacing get() = scrollState.layoutInfo.mainAxisItemSpacing
-}
-
-internal class SliderAdapter internal constructor(
-  val adapter: ScrollAreaState,
-  private val trackSize: Int,
-  private val minHeight: Float,
-  private val reverseLayout: Boolean,
-  private val isVertical: Boolean,
-  private val coroutineScope: CoroutineScope,
-) {
-  private val contentSize get() = adapter.contentSize
-  private val visiblePart: Double
-    get() {
-      val contentSize = contentSize
-      return if (contentSize == 0.0) {
-        1.0
-      } else {
-        (adapter.viewportSize / contentSize).coerceAtMost(1.0)
-      }
-    }
-
-  val thumbSize
-    get() = (trackSize * visiblePart).coerceAtLeast(minHeight.toDouble())
-
-  private val scrollScale: Double
-    get() {
-      val extraScrollbarSpace = trackSize - thumbSize
-      val extraContentSpace = adapter.maxScrollOffset
-      return if (extraContentSpace == 0.0) 1.0 else extraScrollbarSpace / extraContentSpace
-    }
-
-  private val rawPosition: Double
-    get() = scrollScale * adapter.scrollOffset
-
-  val position: Double
-    get() = if (reverseLayout) trackSize - thumbSize - rawPosition else rawPosition
-
-  val bounds get() = position..position + thumbSize
-  private var unscrolledDragDistance = 0.0
-
-  fun onDragStarted() {
-    unscrolledDragDistance = 0.0
-  }
-
-  private suspend fun setPosition(value: Double) {
-    val rawPosition = if (reverseLayout) {
-      trackSize - thumbSize - value
-    } else {
-      value
-    }
-    adapter.scrollTo(rawPosition / scrollScale)
-  }
-
-  private val dragMutex = Mutex()
-
-  fun onDragDelta(offset: Offset) {
-    coroutineScope.launch(start = CoroutineStart.UNDISPATCHED) {
-      dragMutex.withLock {
-        val dragDelta = if (isVertical) offset.y else offset.x
-        val maxScrollPosition = adapter.maxScrollOffset * scrollScale
-        val currentPosition = position
-        val targetPosition = (currentPosition + dragDelta + unscrolledDragDistance).coerceIn(
-          0.0,
-          maxScrollPosition,
-        )
-        val sliderDelta = targetPosition - currentPosition
-        val newPos = position + sliderDelta
-        setPosition(newPos)
-        unscrolledDragDistance += dragDelta - sliderDelta
-      }
-    }
-  }
 }
