+++ settings.gradle.kts
@@ -64,6 +64,7 @@ include(":composeunstyled-stack")
 include(":composeunstyled-focus-ring")
 include(":composeunstyled-primitives")
 include(":composeunstyled-scroll-area")
+include(":composeunstyled-scrollbars")
 include(":composeunstyled-overlay")
 include(":composeunstyled-theming")
 include(":composeunstyled-colored-indication")
